+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Added `ViewModel#view.renderError(container, reason, config)` method to render a render error
 - Fixed the "Show the rest" button on lists to correctly render all remaining elements, instead of just the next N. This regression was introduced in v1.0.0-beta.95.  
 - Updated the `ViewModel#view.renderList()` method to accept the `limit` as a separate argument outside of the `options` object:  
   `renderList(container, itemConfig, data, context, offset, limit, options)`.